add ConnectionManager
